requirements.txt
Django==2.2.12
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
